feat: Add GitHub Actions workflow for build and deploy
This commit adds a GitHub Actions workflow file `cd-demo.yaml` that builds and deploys the application. The workflow is triggered on a release and also supports manual triggering using workflow_dispatch. The workflow runs on Ubuntu using Node.js version 16. It checks out the code, installs dependencies, builds the project, and deploys it using the `github-pages-deploy-action` action.

chore: Update .gitignore

This commit updates the .gitignore file to include the pattern ".env.*" and exclude the file ".env.example". This helps in properly ignoring environment-specific configuration files while including a template for reference.

refactor: Improve code in RestoreSnapshot.vue

This commit improves the code in the RestoreSnapshot.vue component. It now sets a unique key for the table rows in the v-for loop to avoid duplication. It also removes unused imports and variables, resulting in a cleaner and more efficient code.

refactor: Update code in VideoCard.vue

This commit updates the code in the VideoCard.vue component. It returns the response directly instead of creating an object URL in the getThumbnailById function, which simplifies the code and improves performance.

refactor: Update Vite configuration

This commit updates the Vite configuration in vite.config.ts. It sets the root and base paths based on the VITE_APP_MODE environment variable, allowing different configurations for different deployment environments. The aliases for '@', 'images', and 'svgs' are also updated to use fileURLToPath.