feat: Add function to get snapshot time range
This commit adds a new function `getSnapshotTimeRange` in the API module to retrieve the earliest and latest time of the specified IPCam's snapshots. The function makes an API call to `/snapshots/range/${imei}` and returns the response data.

The `TimeRange` interface is also defined to represent the earliest and latest created times of the snapshots.

In the `SnapShotView` component, the dropdown menu now displays the start and end times of the selected IPCam's snapshots. The `selectDate` input field is updated with the last created time if it exists, otherwise it defaults to the current date.