fix: Update data type for IMEI in Snapshot interface
In the `index.ts` file, the data type of the `imei` property in the `Snapshot` interface was changed from `number` to `string`.

---

style: Comment out unused table columns in RestoreSnapshot.vue

In the `RestoreSnapshot.vue` file, several table columns were commented out. These columns were not being used and were unnecessary.

---

feat: Implement IPCam dropdown in RestoreSnapshotView

In the `RestoreSnapshotView.vue` file, an IPCam dropdown component was added. This allows the user to select an IPCam for filtering the snapshots displayed in the RestoreSnapshot component.