refactor(videomaker): set default start and end time
This commit updates the VideoMaker component to set a default start time of '00:00' and end time of '00:00' for the form inputs. Previously, the start and end time were empty strings, but this caused issues with some operations. Now, users will have a default value already filled in when creating a new video.