Fix bug with selectedIPCam not triggering re-fetch of videoList
This commit fixes a bug where the selectedIPCam variable was not triggering a re-fetch of the videoList. With this fix, whenever selectedIPCam changes, the videoList will be re-fetched. Additionally, the getTimeRange and handleDate functions are now being called instead of the resetDate function when selectedIPCam changes. This ensures that the date range is correctly updated based on the new selectedIPCam.